requirements.txt
alembic==1.3.3
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.3.0
Automat==20.2.0
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorclass==2.2.0
constantly==15.1.0
cryptography==2.8
docopt==0.6.2
elasticsearch==7.5.1
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.16.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
more-itertools==8.2.0
mpesa-py==1.0
packaging==20.1
pbr==3.1.1
pip-upgrader==1.4.15
pluggy==0.6.0
psycopg2==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyfcm==1.4.7
PyHamcrest==2.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==3.6.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.20.0
service-identity==18.1.0
six==1.11.0
SQLAlchemy==1.3.13
terminaltables==3.1.0
treq==18.6.0
Twisted==19.10.0
txFCM==0.0.1
urllib3==1.24.2
Werkzeug==0.16.1
WTForms==2.2.1
zope.interface==4.7.1
